pyproject.toml
[tool.poetry]
name = "position5"
version = "0.1.0"
description = "Discord Bot"
authors = ["Arpit Choudhary <arpitkumar147@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
tinydb = "^4.4.0"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
aiohttp = "^3.7.4"
black = "^20.8b1"
cricapi = "^1.2"
"discord.py" = "^1.6.0"
pylint = "^2.7.2"
python-dotenv = "^0.15.0"
requests = "^2.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
skip-string-normalization = false
